Histogram {modelsummary}	R Documentation

datasummary statistic shortcut

Description

This function uses Unicode characters to create a histogram. This can sometimes be useful, but is generally discouraged. Unicode characters can only display a limited number of heights for bars, and the accuracy of output is highly dependent on the platform (typeface, output type, windows vs. mac, etc.). We recommend you use the kableExtra::spec_hist function instead.

Usage

Histogram(x, bins = 10)

Arguments

x
variable to summarize

bins
number of histogram bars

[Package modelsummary version 1.4.3 Index]