Internal function to build table with `gt`

Internal function to build table with `gt`

build_kableExtra(tab, title, stars, stars_note, notes, filename, output, ...)

Arguments

title
string

stars
to indicate statistical significance

FALSE (default): no significance stars.

TRUE: *=.1, **=.05, ***=.01

Named numeric vector for custom stars such as `c('*' = .1, '+' = .05)`

stars_note
passed by `modelsummary()`

notes
list of notes to append to the bottom of the table.

filename
This argument was deprecated in favor of the `output` argument.

output
filename or object type (string)

Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "gt", "html", "markdown", "latex".

When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kable::save_kable`, or `cat` functions.

...
all other arguments are passed to the `tidy` method used to extract estimates from the model. For example, this allows users to set `exponentiate=TRUE` to exponentiate logistic regression coefficients.

Value

tbl_gt object