Balance table: Summary statistics for different subsets of the data (e.g., control and treatment groups)

Balance table: Summary statistics for different subsets of the data (e.g., control and treatment groups)

datasummary_balance(
  formula,
  data,
  output = "default",
  fmt = "%.1f",
  title = NULL,
  notes = NULL,
  align = NULL,
  add_columns = NULL,
  dinm = TRUE,
  dinm_statistic = "std.error",
  ...
)

Arguments

formula
a one-sided formula with the "condition" or "column" variable on the right-hand side.

data
A data.frame (or tibble). If this data includes columns called "blocks", "clusters", and/or "weights", the 'estimatr' package will consider them when calculating the difference in means.

output
filename or object type (string)

Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".

When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kable::save_kable`, or `cat` functions.

fmt
string which specifies how numeric values will be rounded. This string is passed to the `sprintf` function. '%.3f' will keep 3 digits after the decimal point with trailing zero. '%.5f' will keep 5 digits. '%.3e' will use exponential notation. See `?sprintf` for more options.

title
string

notes
list or vector of notes to append to the bottom of the table.

align
A character string of length equal to the number of columns in the table. "lcr" means that the first column will be left-aligned, the 2nd column center-aligned, and the 3rd column right-aligned.

add_columns
a data.frame (or tibble) with the same number of rows as your main table.

dinm
TRUE calculates a difference in means with uncertainty estimates

dinm_statistic
string: "std.error" or "p.value"

...
all other arguments are passed to the `tidy` method used to extract estimates from the model. For example, this allows users to set `exponentiate=TRUE` to exponentiate logistic regression coefficients.

Examples

datasummary_balance(~am, mtcars)

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #rewkadgwxb .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #rewkadgwxb .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #rewkadgwxb .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #rewkadgwxb .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #rewkadgwxb .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #rewkadgwxb .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #rewkadgwxb .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #rewkadgwxb .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #rewkadgwxb .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #rewkadgwxb .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #rewkadgwxb .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #rewkadgwxb .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #rewkadgwxb .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #rewkadgwxb .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #rewkadgwxb .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #rewkadgwxb .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #rewkadgwxb .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #rewkadgwxb .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #rewkadgwxb .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #rewkadgwxb .gt_left {
#>   text-align: left;
#> }
#> 
#> #rewkadgwxb .gt_center {
#>   text-align: center;
#> }
#> 
#> #rewkadgwxb .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #rewkadgwxb .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #rewkadgwxb .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #rewkadgwxb .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #rewkadgwxb .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #rewkadgwxb .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="rewkadgwxb" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_center gt_columns_bottom_border" rowspan="2" colspan="1"> </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2">
#>         <span class="gt_column_spanner">0 (N=19)</span>
#>       </th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2">
#>         <span class="gt_column_spanner">1 (N=13)</span>
#>       </th>
#>       <th class="gt_col_heading gt_center gt_columns_bottom_border" rowspan="2" colspan="1">Diff. in Means</th>
#>       <th class="gt_col_heading gt_center gt_columns_bottom_border" rowspan="2" colspan="1">Std. Error</th>
#>     </tr>
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1">Mean</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1">Std. Dev.</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1">Mean </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1">Std. Dev. </th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr>
#>       <td class="gt_row gt_left">mpg</td>
#>       <td class="gt_row gt_right">17.1</td>
#>       <td class="gt_row gt_right">3.8</td>
#>       <td class="gt_row gt_right">24.4</td>
#>       <td class="gt_row gt_right">6.2</td>
#>       <td class="gt_row gt_right">7.2</td>
#>       <td class="gt_row gt_right">1.9</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">cyl</td>
#>       <td class="gt_row gt_right">6.9</td>
#>       <td class="gt_row gt_right">1.5</td>
#>       <td class="gt_row gt_right">5.1</td>
#>       <td class="gt_row gt_right">1.6</td>
#>       <td class="gt_row gt_right">-1.9</td>
#>       <td class="gt_row gt_right">0.6</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">disp</td>
#>       <td class="gt_row gt_right">290.4</td>
#>       <td class="gt_row gt_right">110.2</td>
#>       <td class="gt_row gt_right">143.5</td>
#>       <td class="gt_row gt_right">87.2</td>
#>       <td class="gt_row gt_right">-146.8</td>
#>       <td class="gt_row gt_right">35.0</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">hp</td>
#>       <td class="gt_row gt_right">160.3</td>
#>       <td class="gt_row gt_right">53.9</td>
#>       <td class="gt_row gt_right">126.8</td>
#>       <td class="gt_row gt_right">84.1</td>
#>       <td class="gt_row gt_right">-33.4</td>
#>       <td class="gt_row gt_right">26.4</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">drat</td>
#>       <td class="gt_row gt_right">3.3</td>
#>       <td class="gt_row gt_right">0.4</td>
#>       <td class="gt_row gt_right">4.0</td>
#>       <td class="gt_row gt_right">0.4</td>
#>       <td class="gt_row gt_right">0.8</td>
#>       <td class="gt_row gt_right">0.1</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">wt</td>
#>       <td class="gt_row gt_right">3.8</td>
#>       <td class="gt_row gt_right">0.8</td>
#>       <td class="gt_row gt_right">2.4</td>
#>       <td class="gt_row gt_right">0.6</td>
#>       <td class="gt_row gt_right">-1.4</td>
#>       <td class="gt_row gt_right">0.2</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">qsec</td>
#>       <td class="gt_row gt_right">18.2</td>
#>       <td class="gt_row gt_right">1.8</td>
#>       <td class="gt_row gt_right">17.4</td>
#>       <td class="gt_row gt_right">1.8</td>
#>       <td class="gt_row gt_right">-0.8</td>
#>       <td class="gt_row gt_right">0.6</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">vs</td>
#>       <td class="gt_row gt_right">0.4</td>
#>       <td class="gt_row gt_right">0.5</td>
#>       <td class="gt_row gt_right">0.5</td>
#>       <td class="gt_row gt_right">0.5</td>
#>       <td class="gt_row gt_right">0.2</td>
#>       <td class="gt_row gt_right">0.2</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">gear</td>
#>       <td class="gt_row gt_right">3.2</td>
#>       <td class="gt_row gt_right">0.4</td>
#>       <td class="gt_row gt_right">4.4</td>
#>       <td class="gt_row gt_right">0.5</td>
#>       <td class="gt_row gt_right">1.2</td>
#>       <td class="gt_row gt_right">0.2</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">carb</td>
#>       <td class="gt_row gt_right">2.7</td>
#>       <td class="gt_row gt_right">1.1</td>
#>       <td class="gt_row gt_right">2.9</td>
#>       <td class="gt_row gt_right">2.2</td>
#>       <td class="gt_row gt_right">0.2</td>
#>       <td class="gt_row gt_right">0.7</td>
#>     </tr>
#>   </tbody>
#>   
#>   
#> </table></div>