Draw a table from a data.frame

Draw a table from a data.frame

datasummary_df(
  data,
  output = "default",
  fmt = 2,
  align = NULL,
  hrule = NULL,
  title = NULL,
  notes = NULL,
  add_rows = NULL,
  add_columns = NULL,
  escape = TRUE,
  ...
)

Arguments

data

A data.frame (or tibble)

output

filename or object type (character string)

Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "modelsummary_list", "gt", "kableExtra", "huxtable", "flextable", "jupyter".

Warning: Users should not supply a file name to the output argument if they intend to customize the table with external packages. See the 'Details' section.

LaTeX compilation requires the booktabs and siunitx packages, but siunitx can be disabled or replaced with global options. See the 'Details' section.

The default output formats and table-making packages can be modified with global options. See the 'Details' section.

fmt

determines how to format numeric values

integer: the number of digits to keep after the period format(round(x, fmt), nsmall=fmt)

character: passed to the sprintf function (e.g., '%.3f' keeps 3 digits with trailing zero). See ?sprintf

function: returns a formatted character string.

LaTeX output: To ensure proper typography, all numeric entries are enclosed in the \num{} command, which requires the siunitx package to be loaded in the LaTeX preamble. This behavior can be altered with global options. See the 'Details' section.

align

A string with a number of characters equal to the number of columns in the table (e.g., align = "lcc"). Valid characters: l, c, r, S.

"l": left-aligned column

"c": centered column

"r": right-aligned column

"d": dot-aligned column. Only supported for LaTeX/PDF tables produced by kableExtra. These commands must appear in the LaTeX preamble (they are added automatically when compiling Rmarkdown documents to PDF):

\usepackage{booktabs}

\usepackage{siunitx}

\newcolumntype{d}{S[input-symbols = ()]}

hrule

position of horizontal rules (integer vector)

title

string

notes

list or vector of notes to append to the bottom of the table.

add_rows

a data.frame (or tibble) with the same number of columns as your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the row positions. See Examples section below.

add_columns

a data.frame (or tibble) with the same number of rows as your main table.

escape

boolean TRUE escapes or substitutes LaTeX/HTML characters which could prevent the file from compiling/displaying. This setting does not affect captions or notes.

...

all other arguments are passed through to the extractor and table-making functions. This allows users to pass arguments directly to modelsummary in order to affect the behavior of other functions behind the scenes. Examples include:

broom::tidy(exponentiate=TRUE) to exponentiate logistic regression. Please see the modelsummary vignette on the package website for important technical notes on this topic.

performance::model_performance(metrics="RMSE") to select goodness-of-fit statistics to extract using the performance package (must have set options(modelsummary_get="easystats") first).