Quick overview of numeric or categorical variables

This function was inspired by the excellent skimr package for R.

datasummary_skim(
  data,
  type = "numeric",
  output = "default",
  fmt = "%.1f",
  histogram = TRUE,
  title = NULL,
  notes = NULL,
  align = NULL,
  escape = TRUE,
  ...
)

Arguments

data
A data.frame (or tibble)

type
of variables to summarize: "numeric" or "categorical" (character)

output
filename or object type (character string)

Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "modelsummary_list", "gt", "kableExtra", "huxtable", "flextable", "jupyter".

Warning: Users should not supply a file name to the output argument if they intend to customize the table with external packages. See the 'Details' section.

LaTeX compilation requires the booktabs and siunitx packages, but siunitx can be disabled or replaced with global options. See the 'Details' section.

The default output formats and table-making packages can be modified with global options. See the 'Details' section.

fmt
determines how to format numeric values

integer: the number of digits to keep after the period format(round(x, fmt), nsmall=fmt)

character: passed to the sprintf function (e.g., '%.3f' keeps 3 digits with trailing zero). See ?sprintf

function: returns a formatted character string.

Note on LaTeX formatting: To ensure proper typography, all numeric entries are enclosed in the \num{} command from the siunitx LaTeX package by default. This behavior can be altered with global options. See the 'Details' section.

histogram
include a histogram (TRUE/FALSE). Supported for:

type = "numeric"

output is "html", "default", "jpg", "png", or "kableExtra"

PDF and HTML documents compiled via Rmarkdown or knitr

See the examples section below for an example of how to use datasummary to include histograms in other formats such as markdown.

title
string

notes
list or vector of notes to append to the bottom of the table.

align
A string with a number of characters equal to the number of columns in the table (e.g., align = "lcc"). Valid characters: l, c, r, S.

"l": left-aligned column

"c": centered column

"r": right-aligned column

"d": dot-aligned column. Only supported for LaTeX/PDF tables produced by kableExtra. These commands must appear in the LaTeX preamble (they are added automatically when compiling Rmarkdown documents to PDF):

\usepackage{booktabs}

\usepackage{siunitx}

\newcolumntype{d}{S[input-open-uncertainty=,input-close-uncertainty=,table-align-text-pre=false]}

escape
boolean TRUE escapes or substitutes LaTeX/HTML characters which could prevent the file from compiling/displaying. This setting does not affect captions or notes.

...
all other arguments are passed through to the table-making functions. This allows users to pass arguments directly to datasummary in order to affect the behavior of other functions behind the scenes, for instance:

kableExtra::kbl(escape=FALSE) to avoid escaping math characters in kableExtra tables.

Examples

if (FALSE) {
dat <- mtcars
dat$vs <- as.logical(dat$vs)
dat$cyl <- as.factor(dat$cyl)
datasummary_skim(dat)
datasummary_skim(dat, "categorical")

# You can use `datasummary` to produce a similar table in different formats.
# Note that the `Histogram` function relies on unicode characters. These
# characters will only display correctly in some operating systems, under some
# locales, using some fonts. Displaying such histograms on Windows computers
# is notoriously tricky. The `modelsummary` authors cannot provide support to
# display these unicode histograms.

f <- All(mtcars) ~ Mean + SD + Min + Median + Max + Histogram
datasummary(f, mtcars, output="markdown")

}