Extract and combine estimates and goodness-of-fit statistics from several statistical models.

Extract and combine estimates and goodness-of-fit statistics from several statistical models.

extract(
  models,
  statistic = "std.error",
  statistic_override = NULL,
  statistic_vertical = TRUE,
  conf_level = 0.95,
  coef_map = NULL,
  coef_omit = NULL,
  gof_map = modelsummary::gof_map,
  gof_omit = NULL,
  add_rows = NULL,
  add_rows_location = NULL,
  stars = FALSE,
  fmt = "%.3f",
  estimate = "estimate",
  ...
)

Arguments

models
a single model object or a (potentially named) list of models to summarize

statistic
string name of the statistic to include in parentheses

Typical values: "conf.int", "std.error", "statistic", "p.value"

Alternative values: any column name produced by `broom::tidy(model)`

statistic_override
manually override the uncertainy estimates. This argument accepts three types of input:

a function or list of functions of length(models) which produce variance-covariance matrices with row and column names equal to the names of your coefficient estimates. For example, `R` supplies the `vcov` function, and the `sandwich` package supplies `vcovHC`, `vcovHAC`, etc.

a list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a list of length(models) vectors with names equal to the names of your coefficient estimates. Numeric vectors are formatted according to `fmt` and placed in brackets, character vectors printed as given.

statistic_vertical
TRUE if statistics should be printed below estimates. FALSE if statistics should be printed beside estimates.

conf_level
confidence level to use for confidence intervals

coef_map
named character vector. Names refer to the original variable names. Values refer to the variable names that will appear in the table. Coefficients which are omitted from this vector will be omitted from the table. The table will be ordered in the same order as this vector.

coef_omit
string regular expression. Omits all matching coefficients from the table (using `stringr::str_detect`).

gof_map
data.frame with four columns: `raw`, `clean`, `fmt`, and `omit`. See `modelsummary::gof_map`

gof_omit
string regular expression. Omits all matching gof statistics from the table (using `stringr::str_detect`).

add_rows
list of character vectors, each of length equal to the number of models + 1.

add_rows_location
integer or NULL. custom rows will be added to the bottom of the table if this parameter is NULL, or after the position set by this integer.

stars
to indicate statistical significance

FALSE (default): no significance stars.

TRUE: *=.1, **=.05, ***=.01

Named numeric vector for custom stars such as `c('*' = .1, '+' = .05)`

fmt
string which specifies how numeric values will be rounded. This string is passed to the `sprintf` function. '%.3f' will keep 3 digits after the decimal point with trailing zero. '%.5f' will keep 5 digits. '%.3e' will use exponential notation. See `?sprintf` for more options.

estimate
character name of the estimate to display. Must be a column name in the dataframe produced by `tidy(model)`. In the vast majority of cases, the default value of this argument should not be changed.

...
all other arguments are passed to the `tidy` method used to extract estimates from the model. For example, this allows users to set `exponentiate=TRUE` to exponentiate logistic regression coefficients.

Value

tibble

Examples

library(modelsummary)
data(trees)
models <- list()
models[['Bivariate']] <- lm(Girth ~ Height, data = trees)
models[['Multivariate']] <- lm(Girth ~ Height + Volume, data = trees)
extract(models)

#> # A tibble: 12 x 5
#>    group     term        statistic    Bivariate Multivariate
#>    <chr>     <chr>       <chr>        <chr>     <chr>       
#>  1 estimates (Intercept) "estimate"   "-6.188"  10.816      
#>  2 estimates (Intercept) "statistic1" "(5.960)" (1.973)     
#>  3 estimates Height      "estimate"   "0.256"   -0.045      
#>  4 estimates Height      "statistic1" "(0.078)" (0.028)     
#>  5 estimates Volume      "estimate"   ""        0.195       
#>  6 estimates Volume      "statistic1" ""        (0.011)     
#>  7 gof       Num.Obs.    ""           "31"      31          
#>  8 gof       R2          ""           "0.270"   0.941       
#>  9 gof       Adj.R2      ""           "0.244"   0.937       
#> 10 gof       AIC         ""           "154.1"   78.2        
#> 11 gof       BIC         ""           "158.4"   84.0        
#> 12 gof       Log.Lik.    ""           "-74.061" -35.116