Factory to create tables in different output formats using standardized inputs.

Factory to create tables in different output formats using standardized inputs.

factory(tab, align = NULL, hrule, notes, output, span, title)

Arguments

tab
table body (data.frame)

align
A character string of length equal to the number of columns in the table. "lcr" means that the first column will be left-aligned, the 2nd column center-aligned, and the 3rd column right-aligned.

hrule
position of horizontal rules (integer vector)

notes
list or vector of notes to append to the bottom of the table.

output
filename or object type (string)

Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".

When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kable::save_kable`, or `cat` functions.

span
list of column label spans

title
string