fmt_decimal {modelsummary}	R Documentation

Rounding with decimal digits in the fmt argument

Description

Rounding with decimal digits in the fmt argument

Usage

fmt_decimal(digits = 3, pdigits = NULL, ...)

Arguments

digits
Number of decimal digits to keep, including trailing zeros.

pdigits
Number of decimal digits to keep for p values. If NULL, the value of digits is used.

...
Additional arguments are passed to the format() function (e.g., big.marks, scientific). See ?format

[Package modelsummary version 1.4.3 Index]