A convenience function to render markdown to html in row and column labels

A convenience function to render markdown to html in row and column labels

fmt_labels_md(tab, position = c("both", "row", "column"))

Arguments

tab
a `gt` table object

position
character string determines wither row, column or both labels should be rendered.

Note

This function only works for HTML output, since the `gt` render tools are less developed for LaTeX and RTF output.