fmt_sci {modelsummary}	R Documentation

Rounding using scientific notation

Description

Rounding using scientific notation

Usage

fmt_sci(digits = 3, ...)

Arguments

digits
a positive integer indicating how many significant digits are to be used for numeric and complex x.

...
additional arguments passed to format().

[Package modelsummary version 1.4.3 Index]