Extract model gof A mostly internal function with some potential uses outside.

Extract model gof A mostly internal function with some potential uses outside.

get_gof(model, vcov_type = NULL, ...)

Arguments

model
a single model object

vcov_type
string vcov type to add at the bottom of the table

...
all other arguments are passed through to the extractor and table-making functions. This allows users to pass arguments directly to modelsummary in order to affect the behavior of other functions behind the scenes. Examples include:

broom::tidy(exponentiate=TRUE) to exponentiate logistic regression. Please see the modelsummary vignette on the package website for important technical notes on this topic.

performance::model_performance(metrics="RMSE") to select goodness-of-fit statistics to extract using the performance package (must have set options(modelsummary_get="easystats") first).