glance_custom_internal.lm {modelsummary}	R Documentation

Avoid namespace conflict when we want to customize glance internally and still allow users to do the same with their own functions

Description

Avoid namespace conflict when we want to customize glance internally and still allow users to do the same with their own functions

Usage

## S3 method for class 'lm'
glance_custom_internal(x, vcov_type = NULL, gof = NULL, ...)

[Package modelsummary version 1.4.3 Index]