map_estimates {modelsummary}	R Documentation

rename and reorder estimates from a single model (before merging to collapse)

Description

rename and reorder estimates from a single model (before merging to collapse)

Usage

map_estimates(estimates, coef_rename, coef_map, coef_omit, group_map)

[Package modelsummary version 1.4.3 Index]