Use the statistic_override function to extract std.error

Use the statistic_override function to extract std.error

statistic_override_function(model, statistic_override)

Arguments

model
object type with an available `tidy` method.

statistic_override
manually override the uncertainy estimates. This argument accepts three types of input:

a function or list of functions of length(models) which produce variance-covariance matrices with row and column names equal to the names of your coefficient estimates. For example, `R` supplies the `vcov` function, and the `sandwich` package supplies `vcovHC`, `vcovHAC`, etc.

a list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a list of length(models) vectors with names equal to the names of your coefficient estimates. Numeric vectors are formatted according to `fmt` and placed in brackets, character vectors printed as given.

Value

tibble