supported_models {modelsummary}	R Documentation

List of model objects from which modelsummary can extract estimates and statistics

Description

List of model objects from which modelsummary can extract estimates and statistics

Usage

supported_models()

[Package modelsummary version 1.4.3 Index]