tidy_custom {modelsummary}	R Documentation

Extract custom information from a model object and turn it into a tidy data.frame or tibble

Description

To customize the output of a model of class lm, you can define a method called tidy_custom.lm which returns a data.frame with a column called "term", and the other columns you want to use as "estimate" or "statistic" in your modelsummary() call. The output of this method must be similar to the result of tidy(model).

Usage

tidy_custom(x)

Arguments

x
An object to be converted into a tidy data.frame or tibble.

Value

A data.frame or tibble with information about model components.

[Package modelsummary version 1.4.3 Index]