Avoid namespace conflict when we want to customize glance internally and still allow users to do the same with their own functions

Avoid namespace conflict when we want to customize glance internally and still allow users to do the same with their own functions

# S3 method for default
tidy_custom_internal(x)