update_modelsummary {modelsummary}	R Documentation

Update modelsummary and its dependencies

Description

Update modelsummary and its dependencies to the latest R-Universe or CRAN versions. The R session needs to be restarted after install.

Usage

update_modelsummary(source = "development")

Arguments

source
one of two strings: "development" or "cran"

[Package modelsummary version 1.4.3 Index]