Balance table: Summary statistics for different subsets of the data (e.g., control and treatment groups)

Balance table: Summary statistics for different subsets of the data (e.g., control and treatment groups)

datasummary_balance(
  formula,
  data,
  output = "default",
  fmt = 1,
  title = NULL,
  notes = NULL,
  align = NULL,
  add_columns = NULL,
  add_rows = NULL,
  dinm = TRUE,
  dinm_statistic = "std.error",
  ...
)

Arguments

formula
a one-sided formula with the "condition" or "column" variable on the right-hand side.

data
A data.frame (or tibble). If this data includes columns called "blocks", "clusters", and/or "weights", the 'estimatr' package will consider them when calculating the difference in means.

output
filename or object type (character string)

Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.

Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".

Warning: the `output` argument cannot be used when customizing tables with external packages. See the 'Details' section below.

fmt
determines how to format numeric values

integer: the number of digits to keep after the period `format(round(x, fmt), nsmall=fmt)`

character: passed to the `sprintf` function (e.g., '%.3f' keeps 3 digits with trailing zero). See `?sprintf`

function: returns a formatted character string.

title
string

notes
list or vector of notes to append to the bottom of the table.

align
A character string of length equal to the number of columns in the table. "lcr" means that the first column will be left-aligned, the 2nd column center-aligned, and the 3rd column right-aligned.

add_columns
a data.frame (or tibble) with the same number of rows as your main table.

add_rows
a data.frame (or tibble) with the same number of columns as your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the row positions. See Examples section below.

dinm
TRUE calculates a difference in means with uncertainty estimates. This option is only available if the `estimatr` package is installed. If `data` includes columns named "blocks", "clusters", or "weights", this information will be taken into account automatically by `estimatr::difference_in_means`.

dinm_statistic
string: "std.error" or "p.value"

...
all other arguments are passed to the `tidy` and `glance` methods used to extract estimates from the model. For example, this allows users to set `exponentiate=TRUE` to exponentiate logistic regression coefficients.

Examples

if (FALSE) {
datasummary_balance(~am, mtcars)
}