Title models with their dependent variables

A convenience function for use with a regression model or list of regression models. Returns a named list of models, where the names are the models' respective dependent variables. Pass your list of models to dvnames before sending to modelsummary to automatically get dependent variable-titled columns.

Usage

dvnames(models, number = FALSE, strip = FALSE, fill = "Model")

Arguments

models

A regression model or list of regression models

number

Should the models be numbered (1), (2), etc., in addition to their dependent variable names?

strip

boolean FALSE returns the dependent variable names as they appear in the model. TRUE returns the dependent variable names as they appear in the data, without transformations.

fill

If insight::find_response() cannot find a response, the column title to use in its place. Set to ' ' to leave blank.

Examples

m1 <- lm(mpg ~ hp, data = mtcars)
m2 <- lm(mpg ~ hp + wt, data = mtcars)

# Without dvnames, column names are (1) and (2)
modelsummary(list(m1, m2))
#> <div id="hifbhlktto" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#hifbhlktto table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #hifbhlktto thead, #hifbhlktto tbody, #hifbhlktto tfoot, #hifbhlktto tr, #hifbhlktto td, #hifbhlktto th {
#>   border-style: none;
#> }
#> 
#> #hifbhlktto p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #hifbhlktto .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #hifbhlktto .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #hifbhlktto .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #hifbhlktto .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #hifbhlktto .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #hifbhlktto .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #hifbhlktto .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #hifbhlktto .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #hifbhlktto .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #hifbhlktto .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #hifbhlktto .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #hifbhlktto .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #hifbhlktto .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #hifbhlktto .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #hifbhlktto .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hifbhlktto .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #hifbhlktto .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #hifbhlktto .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #hifbhlktto .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hifbhlktto .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #hifbhlktto .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hifbhlktto .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #hifbhlktto .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hifbhlktto .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hifbhlktto .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hifbhlktto .gt_left {
#>   text-align: left;
#> }
#> 
#> #hifbhlktto .gt_center {
#>   text-align: center;
#> }
#> 
#> #hifbhlktto .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #hifbhlktto .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #hifbhlktto .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #hifbhlktto .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #hifbhlktto .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #hifbhlktto .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #hifbhlktto .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #hifbhlktto .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #hifbhlktto .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #hifbhlktto .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #hifbhlktto .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #hifbhlktto .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(2)">(2)</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="" class="gt_row gt_left">(Intercept)</td>
#> <td headers="(1)" class="gt_row gt_center">30.099</td>
#> <td headers="(2)" class="gt_row gt_center">37.227</td></tr>
#>     <tr><td headers="" class="gt_row gt_left"></td>
#> <td headers="(1)" class="gt_row gt_center">(1.634)</td>
#> <td headers="(2)" class="gt_row gt_center">(1.599)</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">hp</td>
#> <td headers="(1)" class="gt_row gt_center">-0.068</td>
#> <td headers="(2)" class="gt_row gt_center">-0.032</td></tr>
#>     <tr><td headers="" class="gt_row gt_left"></td>
#> <td headers="(1)" class="gt_row gt_center">(0.010)</td>
#> <td headers="(2)" class="gt_row gt_center">(0.009)</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">wt</td>
#> <td headers="(1)" class="gt_row gt_center"></td>
#> <td headers="(2)" class="gt_row gt_center">-3.878</td></tr>
#>     <tr><td headers="" class="gt_row gt_left" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: #000000;"></td>
#> <td headers="(1)" class="gt_row gt_center" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: #000000;"></td>
#> <td headers="(2)" class="gt_row gt_center" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: #000000;">(0.633)</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">Num.Obs.</td>
#> <td headers="(1)" class="gt_row gt_center">32</td>
#> <td headers="(2)" class="gt_row gt_center">32</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">R2</td>
#> <td headers="(1)" class="gt_row gt_center">0.602</td>
#> <td headers="(2)" class="gt_row gt_center">0.827</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">R2 Adj.</td>
#> <td headers="(1)" class="gt_row gt_center">0.589</td>
#> <td headers="(2)" class="gt_row gt_center">0.815</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">AIC</td>
#> <td headers="(1)" class="gt_row gt_center">181.2</td>
#> <td headers="(2)" class="gt_row gt_center">156.7</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">BIC</td>
#> <td headers="(1)" class="gt_row gt_center">185.6</td>
#> <td headers="(2)" class="gt_row gt_center">162.5</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">Log.Lik.</td>
#> <td headers="(1)" class="gt_row gt_center">-87.619</td>
#> <td headers="(2)" class="gt_row gt_center">-74.326</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">F</td>
#> <td headers="(1)" class="gt_row gt_center">45.460</td>
#> <td headers="(2)" class="gt_row gt_center">69.211</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">RMSE</td>
#> <td headers="(1)" class="gt_row gt_center">3.74</td>
#> <td headers="(2)" class="gt_row gt_center">2.47</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# With dvnames, they are "mpg" and "mpg"
modelsummary(dvnames(list(m1,m2)))
#> <div id="hsiokabywx" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#hsiokabywx table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #hsiokabywx thead, #hsiokabywx tbody, #hsiokabywx tfoot, #hsiokabywx tr, #hsiokabywx td, #hsiokabywx th {
#>   border-style: none;
#> }
#> 
#> #hsiokabywx p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #hsiokabywx .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #hsiokabywx .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #hsiokabywx .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #hsiokabywx .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #hsiokabywx .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #hsiokabywx .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #hsiokabywx .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #hsiokabywx .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #hsiokabywx .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #hsiokabywx .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #hsiokabywx .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #hsiokabywx .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #hsiokabywx .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #hsiokabywx .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #hsiokabywx .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hsiokabywx .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #hsiokabywx .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #hsiokabywx .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #hsiokabywx .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hsiokabywx .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #hsiokabywx .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hsiokabywx .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #hsiokabywx .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hsiokabywx .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hsiokabywx .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hsiokabywx .gt_left {
#>   text-align: left;
#> }
#> 
#> #hsiokabywx .gt_center {
#>   text-align: center;
#> }
#> 
#> #hsiokabywx .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #hsiokabywx .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #hsiokabywx .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #hsiokabywx .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #hsiokabywx .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #hsiokabywx .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #hsiokabywx .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #hsiokabywx .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #hsiokabywx .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #hsiokabywx .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #hsiokabywx .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #hsiokabywx .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="mpg">mpg</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="mpg ">mpg </th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="" class="gt_row gt_left">(Intercept)</td>
#> <td headers="mpg" class="gt_row gt_center">30.099</td>
#> <td headers="mpg" class="gt_row gt_center">37.227</td></tr>
#>     <tr><td headers="" class="gt_row gt_left"></td>
#> <td headers="mpg" class="gt_row gt_center">(1.634)</td>
#> <td headers="mpg" class="gt_row gt_center">(1.599)</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">hp</td>
#> <td headers="mpg" class="gt_row gt_center">-0.068</td>
#> <td headers="mpg" class="gt_row gt_center">-0.032</td></tr>
#>     <tr><td headers="" class="gt_row gt_left"></td>
#> <td headers="mpg" class="gt_row gt_center">(0.010)</td>
#> <td headers="mpg" class="gt_row gt_center">(0.009)</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">wt</td>
#> <td headers="mpg" class="gt_row gt_center"></td>
#> <td headers="mpg" class="gt_row gt_center">-3.878</td></tr>
#>     <tr><td headers="" class="gt_row gt_left" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: #000000;"></td>
#> <td headers="mpg" class="gt_row gt_center" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: #000000;"></td>
#> <td headers="mpg" class="gt_row gt_center" style="border-bottom-width: 1px; border-bottom-style: solid; border-bottom-color: #000000;">(0.633)</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">Num.Obs.</td>
#> <td headers="mpg" class="gt_row gt_center">32</td>
#> <td headers="mpg" class="gt_row gt_center">32</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">R2</td>
#> <td headers="mpg" class="gt_row gt_center">0.602</td>
#> <td headers="mpg" class="gt_row gt_center">0.827</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">R2 Adj.</td>
#> <td headers="mpg" class="gt_row gt_center">0.589</td>
#> <td headers="mpg" class="gt_row gt_center">0.815</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">AIC</td>
#> <td headers="mpg" class="gt_row gt_center">181.2</td>
#> <td headers="mpg" class="gt_row gt_center">156.7</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">BIC</td>
#> <td headers="mpg" class="gt_row gt_center">185.6</td>
#> <td headers="mpg" class="gt_row gt_center">162.5</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">Log.Lik.</td>
#> <td headers="mpg" class="gt_row gt_center">-87.619</td>
#> <td headers="mpg" class="gt_row gt_center">-74.326</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">F</td>
#> <td headers="mpg" class="gt_row gt_center">45.460</td>
#> <td headers="mpg" class="gt_row gt_center">69.211</td></tr>
#>     <tr><td headers="" class="gt_row gt_left">RMSE</td>
#> <td headers="mpg" class="gt_row gt_center">3.74</td>
#> <td headers="mpg" class="gt_row gt_center">2.47</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>