Title models with their dependent variables

A convenience function for use with a regression model or list of regression models. Returns a named list of models, where the names are the models' respective dependent variables. Pass your list of models to dvnames before sending to modelsummary to automatically get dependent variable-titled columns.

Usage

dvnames(models, number = FALSE, strip = FALSE, fill = "Model")

Arguments

models

A regression model or list of regression models

number

Should the models be numbered (1), (2), etc., in addition to their dependent variable names?

strip

boolean FALSE returns the dependent variable names as they appear in the model. TRUE returns the dependent variable names as they appear in the data, without transformations.

fill

If insight::find_response() cannot find a response, the column title to use in its place. Set to ' ' to leave blank.

Examples

m1 <- lm(mpg ~ hp, data = mtcars)
m2 <- lm(mpg ~ hp + wt, data = mtcars)

# Without dvnames, column names are (1) and (2)
modelsummary(list(m1, m2))
#> <table class="table" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;">   </th>
#>    <th style="text-align:center;">  (1) </th>
#>    <th style="text-align:center;">   (2) </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> (Intercept) </td>
#>    <td style="text-align:center;"> 30.099 </td>
#>    <td style="text-align:center;"> 37.227 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:center;"> (1.634) </td>
#>    <td style="text-align:center;"> (1.599) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> hp </td>
#>    <td style="text-align:center;"> −0.068 </td>
#>    <td style="text-align:center;"> −0.032 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:center;"> (0.010) </td>
#>    <td style="text-align:center;"> (0.009) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> wt </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;"> −3.878 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;box-shadow: 0px 1.5px">  </td>
#>    <td style="text-align:center;box-shadow: 0px 1.5px">  </td>
#>    <td style="text-align:center;box-shadow: 0px 1.5px"> (0.633) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Num.Obs. </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 32 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> R2 </td>
#>    <td style="text-align:center;"> 0.602 </td>
#>    <td style="text-align:center;"> 0.827 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> R2 Adj. </td>
#>    <td style="text-align:center;"> 0.589 </td>
#>    <td style="text-align:center;"> 0.815 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AIC </td>
#>    <td style="text-align:center;"> 181.2 </td>
#>    <td style="text-align:center;"> 156.7 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BIC </td>
#>    <td style="text-align:center;"> 185.6 </td>
#>    <td style="text-align:center;"> 162.5 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Log.Lik. </td>
#>    <td style="text-align:center;"> −87.619 </td>
#>    <td style="text-align:center;"> −74.326 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> F </td>
#>    <td style="text-align:center;"> 45.460 </td>
#>    <td style="text-align:center;"> 69.211 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> RMSE </td>
#>    <td style="text-align:center;"> 3.74 </td>
#>    <td style="text-align:center;"> 2.47 </td>
#>   </tr>
#> </tbody>
#> </table>

# With dvnames, they are "mpg" and "mpg"
modelsummary(dvnames(list(m1,m2)))
#> <table class="table" style="width: auto !important; margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;">   </th>
#>    <th style="text-align:center;"> mpg </th>
#>    <th style="text-align:center;"> mpg  </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> (Intercept) </td>
#>    <td style="text-align:center;"> 30.099 </td>
#>    <td style="text-align:center;"> 37.227 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:center;"> (1.634) </td>
#>    <td style="text-align:center;"> (1.599) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> hp </td>
#>    <td style="text-align:center;"> −0.068 </td>
#>    <td style="text-align:center;"> −0.032 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;">  </td>
#>    <td style="text-align:center;"> (0.010) </td>
#>    <td style="text-align:center;"> (0.009) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> wt </td>
#>    <td style="text-align:center;">  </td>
#>    <td style="text-align:center;"> −3.878 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;box-shadow: 0px 1.5px">  </td>
#>    <td style="text-align:center;box-shadow: 0px 1.5px">  </td>
#>    <td style="text-align:center;box-shadow: 0px 1.5px"> (0.633) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Num.Obs. </td>
#>    <td style="text-align:center;"> 32 </td>
#>    <td style="text-align:center;"> 32 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> R2 </td>
#>    <td style="text-align:center;"> 0.602 </td>
#>    <td style="text-align:center;"> 0.827 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> R2 Adj. </td>
#>    <td style="text-align:center;"> 0.589 </td>
#>    <td style="text-align:center;"> 0.815 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> AIC </td>
#>    <td style="text-align:center;"> 181.2 </td>
#>    <td style="text-align:center;"> 156.7 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> BIC </td>
#>    <td style="text-align:center;"> 185.6 </td>
#>    <td style="text-align:center;"> 162.5 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Log.Lik. </td>
#>    <td style="text-align:center;"> −87.619 </td>
#>    <td style="text-align:center;"> −74.326 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> F </td>
#>    <td style="text-align:center;"> 45.460 </td>
#>    <td style="text-align:center;"> 69.211 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> RMSE </td>
#>    <td style="text-align:center;"> 3.74 </td>
#>    <td style="text-align:center;"> 2.47 </td>
#>   </tr>
#> </tbody>
#> </table>