Extract model estimates. A mostly internal function with some potential uses outside.

Extract model estimates. A mostly internal function with some potential uses outside.

get_estimates(model, conf_level = 0.95, vcov = NULL, ...)

Arguments

model
a single model object

conf_level
confidence level to use for confidence intervals

vcov
robust standard errors and other manual statistics. The vcov argument accepts five types of input (see the 'Details' and 'Examples' sections below):

string, vector, or list of strings: "robust", "HC", "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5", "stata", or "classical" (alias "constant" or "iid").

formula or list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).

function or list of functions which return variance-covariance matrices with row and column names equal to the names of your coefficient estimates (e.g., stats::vcov, sandwich::vcovHC).

list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below.

...
all other arguments are passed through to the extractor and table-making functions. This allows users to pass arguments directly to modelsummary in order to affect the behavior of other functions behind the scenes. Examples include:

broom::tidy(exponentiate=TRUE) to exponentiate logistic regression

kableExtra::kbl(escape=FALSE) to avoid escaping math characters in kableExtra tables.

performance::model_performance(metrics="RMSE") to select goodness-of-fit statistics to extract using the performance package (must have set options(modelsummary_get="easystats") first).