get_gof {modelsummary}	R Documentation

Extract model gof A mostly internal function with some potential uses outside.

Description

Extract model gof A mostly internal function with some potential uses outside.

Usage

get_gof(model, vcov_type = NULL, ...)

Arguments

model
a single model object

vcov_type
string vcov type to add at the bottom of the table

...
all other arguments are passed through to three functions. See the documentation of these functions for lists of available arguments.

parameters::model_parameters extracts parameter estimates. Available arguments depend on model type, but include:

standardize, centrality, dispersion, test, ci_method, prior, diagnostic, rope_range, power, cluster, etc.

performance::model_performance extracts goodness-of-fit statistics. Available arguments depend on model type, but include:

metrics, estimator, etc.

kableExtra::kbl or gt::gt draw tables, depending on the value of the output argument.

[Package modelsummary version 1.4.3 Index]