glance_custom {modelsummary}	R Documentation

Extract custom information from a model object and turn it into a tidy data.frame or tibble with a single row.

Description

To customize the output of a model of class lm, you can define a new method called glance_custom.lm which returns a one-row data.frame.

Usage

glance_custom(x, ...)

Arguments

x
model or other R object to convert to single-row data frame

...
ellipsis

[Package modelsummary version 1.4.3 Index]