Extract custom information from a model object and turn it into a tidy data.frame or tibble with a single row.

To customize the output of a model of class lm, you can define a new method called glance_custom.lm which returns a one-row data.frame.

Usage

glance_custom(x, ...)

Arguments

x

model or other R object to convert to single-row data frame

...

ellipsis

Methods

See the following help topics for more details about individual methods:broom

data.frame_tidiers: data.frame

durbinWatsonTest_tidiers: durbinWatsonTest

glance_gam_hastie: Gam

glance.aareg: aareg

glance.anova: anova

glance.aov: aov

glance.Arima: Arima

glance.betamfx: betamfx

glance.betareg: betareg

glance.biglm: biglm

glance.binDesign: binDesign

glance.cch: cch

glance.clm: clm

glance.clmm: clmm

glance.coeftest: coeftest

glance.coxph: coxph

glance.crr: crr

glance.cv.glmnet: cv.glmnet

glance.drc: drc

glance.ergm: ergm

glance.factanal: factanal

glance.felm: felm

glance.fitdistr: fitdistr

glance.fixest: fixest

glance.gam: gam

glance.garch: garch

glance.geeglm: geeglm

glance.glm: glm

glance.glmnet: glmnet

glance.glmRob: glmRob

glance.gmm: gmm

glance.ivreg: ivreg

glance.kmeans: kmeans

glance.lavaan: lavaan

glance.lm: lm

glance.lmodel2: lmodel2

glance.lmRob: lmRob

glance.margins: margins

glance.Mclust: Mclust

glance.mfx: logitmfx, mfx, negbinmfx, poissonmfx, probitmfx

glance.mjoint: mjoint

glance.mlogit: mlogit

glance.muhaz: muhaz

glance.multinom: multinom

glance.negbin: negbin

glance.nlrq: nlrq

glance.nls: nls

glance.orcutt: orcutt

glance.pam: pam

glance.plm: plm

glance.poLCA: poLCA

glance.polr: polr

glance.pyears: pyears

glance.ridgelm: ridgelm

glance.rlm: rlm

glance.rma: rma

glance.robustbase.lmrob: lmrob

glance.rq: rq

glance.sarlm: sarlm

glance.smooth.spline: smooth.spline

glance.speedglm: speedglm

glance.speedlm: speedlm

glance.summary.lm: summary.lm

glance.survdiff: survdiff

glance.survexp: survexp

glance.survfit: survfit

glance.survreg: survreg

glance.svyglm: svyglm

glance.svyolr: svyolr

glance.varest: varest

list_tidiers: list

null_tidiers: NULL

summary_tidiers: summaryDefault

tidy.htest: htest

estimatr

estimatr_glancers: difference_in_means, horvitz_thompson, iv_robust, lh_robust, lm_robust