gof_map {modelsummary}	R Documentation

Data.frame used to clean up and format goodness-of-fit statistics

Description

By default, this data frame is passed to the 'gof_map' argument of the 'modelsummary' function. Users can modify this data frame to customize the list of statistics to display and their format. See example below.

Usage

gof_map

Format

data.frame with 4 columns of character data: raw, clean, fmt, omit

Examples

if (identical(Sys.getenv("pkgdown"),  "true")) {

library(modelsummary)
mod <- lm(wt ~ drat, data = mtcars)
gm <- modelsummary::gof_map
gm$omit[gm$raw == 'deviance'] <- FALSE
gm$fmt[gm$raw == 'r.squared'] <- "%.5f"
modelsummary(mod, gof_map = gm)
}

[Package modelsummary version 1.4.3 Index]