msummary() is a shortcut to modelsummary()

Create beautiful and customizable tables to summarize several statistical models side-by-side. This function supports dozens of statistical models, and it can produce tables in HTML, LaTeX, Word, Markdown, PDF, PowerPoint, Excel, RTF, JPG, or PNG. The appearance of the tables can be customized extensively by specifying the output argument, and by using functions from one of the supported table customization packages: kableExtra, gt, flextable, huxtable, DT. For more information, see the Details and Examples sections below, and the vignettes on the modelsummary website: https://vincentarelbundock.github.io/modelsummary/

The modelsummary Vignette includes dozens of examples of tables with extensive customizations.

The Appearance Vignette shows how to modify the look of tables.

Usage

msummary(
  models,
  output = "default",
  fmt = 3,
  estimate = "estimate",
  statistic = "std.error",
  vcov = NULL,
  conf_level = 0.95,
  exponentiate = FALSE,
  stars = FALSE,
  shape = term + statistic ~ model,
  coef_map = NULL,
  coef_omit = NULL,
  coef_rename = FALSE,
  gof_map = NULL,
  gof_omit = NULL,
  group_map = NULL,
  add_columns = NULL,
  add_rows = NULL,
  align = NULL,
  notes = NULL,
  title = NULL,
  escape = TRUE,
  ...
)

Arguments

models

a model, (named) list of models, or nested list of models.

Single model: modelsummary(model)

Unnamed list of models: modelsummary(list(model1, model2))

Models are labelled automatically. The default label style can be altered by setting a global option. See below.

Named list of models: modelsummary(list("A"=model1, "B"=model2))

Models are labelled using the list names.

Nested list of models: When using the shape="rbind" argument, models can be a nested list of models to display "panels" or "stacks" of regression models. See the shape argument documentation and examples below.

output

filename or object type (character string)

Supported filename extensions: .docx, .html, .tex, .md, .txt, .csv, .xlsx, .png, .jpg

Supported object types: "default", "html", "markdown", "latex", "latex_tabular", "data.frame", "gt", "kableExtra", "huxtable", "flextable", "DT", "jupyter". The "modelsummary_list" value produces a lightweight object which can be saved and fed back to the modelsummary function.

The "default" output format can be set to "kableExtra", "gt", "flextable", "huxtable", "DT", or "markdown"

If the user does not choose a default value, the packages listed above are tried in sequence.

Session-specific configuration: options("modelsummary_factory_default" = "gt")

Persistent configuration: config_modelsummary(output = "markdown")

Warning: Users should not supply a file name to the output argument if they intend to customize the table with external packages. See the 'Details' section.

LaTeX compilation requires the booktabs and siunitx packages, but siunitx can be disabled or replaced with global options. See the 'Details' section.

fmt

how to format numeric values: integer, user-supplied function, or modelsummary function.

Integer: Number of decimal digits

User-supplied functions:

Any function which accepts a numeric vector and returns a character vector of the same length.

modelsummary functions:

fmt = fmt_significant(2): Two significant digits (at the term-level)

fmt = fmt_decimal(digits = 2, pdigits = 3): Decimal digits for estimate and p values

fmt = fmt_sprintf("%.3f"): See ?sprintf

fmt = fmt_term("(Intercept)" = 1, "X" = 2): Format terms differently

fmt = fmt_statistic("estimate" = 1, "r.sqared" = 6): Format statistics differently.

fmt = fmt_identity(): unformatted raw values

string:

Note on LaTeX output: To ensure proper typography, all numeric entries are enclosed in the \num{} command, which requires the siunitx package to be loaded in the LaTeX preamble. This behavior can be altered with global options. See the 'Details' section.

estimate

a single string or a character vector of length equal to the number of models. Valid entries include any column name of the data.frame produced by get_estimates(model), and strings with curly braces compatible with the glue package format. Examples:

"estimate"

"{estimate} ({std.error}){stars}"

"{estimate} [{conf.low}, {conf.high}]"

statistic

vector of strings or glue strings which select uncertainty statistics to report vertically below the estimate. NULL omits all uncertainty statistics.

"conf.int", "std.error", "statistic", "p.value", "conf.low", "conf.high", . or any column name produced by get_estimates(model)

glue package strings with braces, with or without R functions, such as:

"{p.value} [{conf.low}, {conf.high}]"

"Std.Error: {std.error}"

`"exp(estimate) * std.error"

Numbers are automatically rounded and converted to strings. To apply functions to their numeric values, as in the last glue example, users must set fmt=NULL.

Parentheses are added automatically unless the string includes glue curly braces {}.

Some statistics are not supported for all models. See column names in get_estimates(model), and visit the website to learn how to add custom statistics.

vcov

robust standard errors and other manual statistics. The vcov argument accepts six types of input (see the 'Details' and 'Examples' sections below):

NULL returns the default uncertainty estimates of the model object

string, vector, or (named) list of strings. "iid", "classical", and "constant" are aliases for NULL, which returns the model's default uncertainty estimates. The strings "HC", "HC0", "HC1" (alias: "stata"), "HC2", "HC3" (alias: "robust"), "HC4", "HC4m", "HC5", "HAC", "NeweyWest", "Andrews", "panel-corrected", "outer-product", and "weave" use variance-covariance matrices computed using functions from the sandwich package, or equivalent method. The behavior of those functions can (and sometimes must) be altered by passing arguments to sandwich directly from modelsummary through the ellipsis (...), but it is safer to define your own custom functions as described in the next bullet.

function or (named) list of functions which return variance-covariance matrices with row and column names equal to the names of your coefficient estimates (e.g., stats::vcov, sandwich::vcovHC, function(x) vcovPC(x, cluster="country")).

formula or (named) list of formulas with the cluster variable(s) on the right-hand side (e.g., ~clusterid).

named list of length(models) variance-covariance matrices with row and column names equal to the names of your coefficient estimates.

a named list of length(models) vectors with names equal to the names of your coefficient estimates. See 'Examples' section below. Warning: since this list of vectors can include arbitrary strings or numbers, modelsummary cannot automatically calculate p values. The stars argument may thus use incorrect significance thresholds when vcov is a list of vectors.

conf_level

numeric value between 0 and 1. confidence level to use for confidence intervals. Setting this argument to NULL does not extract confidence intervals, which can be faster for some models.

exponentiate

TRUE, FALSE, or logical vector of length equal to the number of models. If TRUE, the estimate, conf.low, and conf.high statistics are exponentiated, and the std.error is transformed to exp(estimate)*std.error.

stars

to indicate statistical significance

FALSE (default): no significance stars.

TRUE: +=.1, *=.05, **=.01, ***=0.001

Named numeric vector for custom stars such as c('*' = .1, '+' = .05)

Note: a legend will not be inserted at the bottom of the table when the estimate or statistic arguments use "glue strings" with {stars}.

shape

NULL, formula, or string which determines the shape of a table.

NULL: Default shape with terms in rows and models in columns.

Formula: The left side determines what appears on rows, and the right side determines what appears on columns. The formula can include one or more group identifier(s) to display related terms together, which can be useful for models with multivariate outcomes or grouped coefficients (See examples section below). The group identifier(s) must be column names produced by: get_estimates(model). The group identifier(s) can be combined with the term identifier in a single column by using the colon to represent an interaction. If an incomplete formula is supplied (e.g., ~statistic), modelsummary tries to complete it automatically. Goodness-of-fit statistics are only appended to the bottom of the table when model is on the right hand side of the formula (i.e., columns). Potential shape values include:

term + statistic ~ model: default

term ~ model + statistic: statistics in separate columns

model + statistic ~ term: models in rows and terms in columns

term + response + statistic ~ model: term and group id in separate columns

term : response + statistic ~ model: term and group id in a single column

term ~ response

String: "rbind" or "rcollapse" to bind rows of two or more regression tables to create "panels" or "stacks" of regression models.

the models argument must be a (potentially named) nested list of models.

Unnamed nested list with 2 panels: list(list(model1, model2), list(model3, model4))

Named nested list with 2 panels: list("Panel A" = list(model1, model2), "Panel B" = list(model3, model4))

Named panels and named models: list("Panel A" = list("(I)" = model1, "(II)" = model2), "Panel B" = list("(I)" = model3, "(II)" = model4))

"rbind": Bind the rows of independent regression tables

"rcollapse": Bind the rows of regression tables and create a panel at the bottom where we "collapse" goodness-of-fit statistics which are identical across models.

coef_map

character vector. Subset, rename, and reorder coefficients. Coefficients omitted from this vector are omitted from the table. The order of the vector determines the order of the table. coef_map can be a named or an unnamed character vector. If coef_map is a named vector, its values define the labels that must appear in the table, and its names identify the original term names stored in the model object: c("hp:mpg"="HPxM/G"). See Examples section below.

coef_omit

integer vector or regular expression to identify which coefficients to omit (or keep) from the table. Positive integers determine which coefficients to omit. Negative integers determine which coefficients to keep. A regular expression can be used to omit coefficients, and perl-compatible "negative lookaheads" can be used to specify which coefficients to keep in the table. Examples:

c(2, 3, 5): omits the second, third, and fifth coefficients.

c(-2, -3, -5): negative values keep the second, third, and fifth coefficients.

"ei": omit coefficients matching the "ei" substring.

"^Volume$": omit the "Volume" coefficient.

"ei|rc": omit coefficients matching either the "ei" or the "rc" substrings.

"^(?!Vol)": keep coefficients starting with "Vol" (inverse match using a negative lookahead).

"^(?!.*ei)": keep coefficients matching the "ei" substring.

"^(?!.*ei|.*pt)": keep coefficients matching either the "ei" or the "pt" substrings.

See the Examples section below for complete code.

coef_rename

logical, named or unnamed character vector, or function

Logical: TRUE renames variables based on the "label" attribute of each column. See the Example section below.

Unnamed character vector of length equal to the number of coefficients in the final table, after coef_omit is applied.

Named character vector: Values refer to the variable names that will appear in the table. Names refer to the original term names stored in the model object. Ex: c("hp:mpg"="hp X mpg")

Function: Accepts a character vector of the model's term names and returns a named vector like the one described above. The modelsummary package supplies a coef_rename() function which can do common cleaning tasks: modelsummary(model, coef_rename = coef_rename)

gof_map

rename, reorder, and omit goodness-of-fit statistics and other model information. This argument accepts 4 types of values:

NULL (default): the modelsummary::gof_map dictionary is used for formatting, and all unknown statistic are included.

character vector: "all", "none", or a vector of statistics such as c("rmse", "nobs", "r.squared"). Elements correspond to colnames in the data.frame produced by get_gof(model). The modelsummary::gof_map default dictionary is used to format and rename statistics.

NA: excludes all statistics from the bottom part of the table.

data.frame with 3 columns named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples' section below.

list of lists, each of which includes 3 elements named "raw", "clean", "fmt". Unknown statistics are omitted. See the 'Examples section below'.

gof_omit

string regular expression (perl-compatible) used to determine which statistics to omit from the bottom section of the table. A "negative lookahead" can be used to specify which statistics to keep in the table. Examples:

"IC": omit statistics matching the "IC" substring.

"BIC|AIC": omit statistics matching the "AIC" or "BIC" substrings.

"^(?!.*IC)": keep statistics matching the "IC" substring.

group_map

named or unnamed character vector. Subset, rename, and reorder coefficient groups specified a grouping variable specified in the shape argument formula. This argument behaves like coef_map.

add_columns

a data.frame (or tibble) with the same number of rows as #' your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the columns positions. See Examples section below.

add_rows

a data.frame (or tibble) with the same number of columns as your main table. By default, rows are appended to the bottom of the table. You can define a "position" attribute of integers to set the row positions. See Examples section below.

align

A string with a number of characters equal to the number of columns in the table (e.g., align = "lcc"). Valid characters: l, c, r, d.

"l": left-aligned column

"c": centered column

"r": right-aligned column

"d": dot-aligned column. For LaTeX/PDF output, this option requires at least version 3.0.25 of the siunitx LaTeX package. These commands must appear in the LaTeX preamble (they are added automatically when compiling Rmarkdown documents to PDF):

\usepackage{booktabs}

\usepackage{siunitx}

\newcolumntype{d}{S[ input-open-uncertainty=, input-close-uncertainty=, parse-numbers = false, table-align-text-pre=false, table-align-text-post=false ]}

notes

list or vector of notes to append to the bottom of the table.

title

string

escape

boolean TRUE escapes or substitutes LaTeX/HTML characters which could prevent the file from compiling/displaying. This setting does not affect captions or notes.

...

all other arguments are passed through to three functions. See the documentation of these functions for lists of available arguments.

parameters::model_parameters extracts parameter estimates. Available arguments depend on model type, but include:

standardize, centrality, dispersion, test, ci_method, prior, diagnostic, rope_range, power, cluster, etc.

performance::model_performance extracts goodness-of-fit statistics. Available arguments depend on model type, but include:

metrics, estimator, etc.

kableExtra::kbl or gt::gt draw tables, depending on the value of the output argument.

Value

a regression table in a format determined by the output argument.

Details

output

The modelsummary_list output is a lightweight format which can be used to save model results, so they can be fed back to modelsummary later to avoid extracting results again.

When a file name with a valid extension is supplied to the output argument, the table is written immediately to file. If you want to customize your table by post-processing it with an external package, you need to choose a different output format and saving mechanism. Unfortunately, the approach differs from package to package:

gt: set output="gt", post-process your table, and use the gt::gtsave function.

kableExtra: set output to your destination format (e.g., "latex", "html", "markdown"), post-process your table, and use kableExtra::save_kable function.

vcov

To use a string such as "robust" or "HC0", your model must be supported by the sandwich package. This includes objects such as: lm, glm, survreg, coxph, mlogit, polr, hurdle, zeroinfl, and more.

NULL, "classical", "iid", and "constant" are aliases which do not modify uncertainty estimates and simply report the default standard errors stored in the model object.

One-sided formulas such as ~clusterid are passed to the sandwich::vcovCL function.

Matrices and functions producing variance-covariance matrices are first passed to lmtest. If this does not work, modelsummary attempts to take the square root of the diagonal to adjust "std.error", but the other uncertainty estimates are not be adjusted.

Numeric vectors are formatted according to fmt and placed in brackets. Character vectors printed as given, without parentheses.

If your model type is supported by the lmtest package, the vcov argument will try to use that package to adjust all the uncertainty estimates, including "std.error", "statistic", "p.value", and "conf.int". If your model is not supported by lmtest, only the "std.error" will be adjusted by, for example, taking the square root of the matrix's diagonal.

Global Options

The behavior of modelsummary can be modified by setting global options. For example:

options(modelsummary_model_labels = "roman")

The rest of this section describes each of the options above.

Model labels: default column names

These global option changes the style of the default column headers:

options(modelsummary_model_labels = "roman")

options(modelsummary_panel_labels = "roman")

The supported styles are: "model", "panel", "arabic", "letters", "roman", "(arabic)", "(letters)", "(roman)""

The panel-specific option is only used when shape="rbind"

Table-making packages

modelsummary supports 4 table-making packages: kableExtra, gt, flextable, huxtable, and DT. Some of these packages have overlapping functionalities. For example, 3 of those packages can export to LaTeX. To change the default backend used for a specific file format, you can use the options function:

options(modelsummary_factory_html = 'kableExtra')

options(modelsummary_factory_latex = 'gt')

options(modelsummary_factory_word = 'huxtable')

options(modelsummary_factory_png = 'gt')

Table themes

Change the look of tables in an automated and replicable way, using the modelsummary theming functionality. See the vignette: https://vincentarelbundock.github.io/modelsummary/articles/appearance.html

modelsummary_theme_gt

modelsummary_theme_kableExtra

modelsummary_theme_huxtable

modelsummary_theme_flextable

modelsummary_theme_dataframe

Model extraction functions

modelsummary can use two sets of packages to extract information from statistical models: the easystats family (performance and parameters) and broom. By default, it uses easystats first and then falls back on broom in case of failure. You can change the order of priorities or include goodness-of-fit extracted by both packages by setting:

options(modelsummary_get = "broom")

options(modelsummary_get = "easystats")

options(modelsummary_get = "all")

Formatting numeric entries

By default, LaTeX tables enclose all numeric entries in the \num{} command from the siunitx package. To prevent this behavior, or to enclose numbers in dollar signs (for LaTeX math mode), users can call:

options(modelsummary_format_numeric_latex = "plain")

options(modelsummary_format_numeric_latex = "mathmode")

A similar option can be used to display numerical entries using MathJax in HTML tables:

options(modelsummary_format_numeric_html = "mathjax")

Parallel computation

It can take a long time to compute and extract summary statistics from certain models (e.g., Bayesian). In those cases, users can parallelize the process. Since parallelization occurs at the model level, no speedup is available for tables with a single model. Users on mac or linux can launch parallel computation using the built-in parallel package. All they need to do is supply a mc.cores argument which will be pushed forward to the parallel::mclapply function:

modelsummary(model_list, mc.cores = 5)

All users can also use the future.apply package to parallelize model summaries. For example, to use 4 cores to extract results:

library(future.apply)
plan(multicore, workers = 4)
options("modelsummary_future" = TRUE)
modelsummary(model_list)

Note that the "multicore" plan only parallelizes under mac or linux. Windows users can use plan(multisession) instead. However, note that the first time modelsummary() is called under multisession can be a fair bit longer, because of extra costs in passing data to and loading required packages on to workers. Subsequent calls to modelsummary() will often be much faster.

Some users have reported difficult to reproduce errors when using the future package with some packages. The future parallelization in modelsummary can be disabled by calling:

options("modelsummary_future" = FALSE)

Examples

library(modelsummary)

# load data and estimate models
utils::data(trees)
models <- list()
models[['Bivariate']] <- lm(Girth ~ Height, data = trees)
models[['Multivariate']] <- lm(Girth ~ Height + Volume, data = trees)

# simple table
modelsummary(models)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Volume 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# statistic
modelsummary(models, statistic = NULL)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

Height 	0.256 	-0.045

Volume 	 	0.195

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models, statistic = 'p.value')

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(0.308) 	(<0.001)

Height 	0.256 	-0.045

(0.003) 	(0.119)

Volume 	 	0.195

(<0.001)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models, statistic = 'statistic')

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(-1.038) 	(5.482)

Height 	0.256 	-0.045

(3.272) 	(-1.609)

Volume 	 	0.195

(17.816)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models, statistic = 'conf.int', conf_level = 0.99)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

[-22.617, 10.240] 	[5.364, 16.269]

Height 	0.256 	-0.045

[0.040, 0.471] 	[-0.124, 0.033]

Volume 	 	0.195

[0.165, 0.225]

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models, statistic = c("t = {statistic}",
                                   "se = {std.error}",
                                   "conf.int"))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

t = -1.038 	t = 5.482

se = 5.960 	se = 1.973

[-18.378, 6.002] 	[6.774, 14.858]

Height 	0.256 	-0.045

t = 3.272 	t = -1.609

se = 0.078 	se = 0.028

[0.096, 0.416] 	[-0.103, 0.012]

Volume 	 	0.195

t = 17.816

se = 0.011

[0.173, 0.218]

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# estimate
modelsummary(models,
  statistic = NULL,
  estimate = "{estimate} [{conf.low}, {conf.high}]")

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 [-18.378, 6.002] 	10.816 [6.774, 14.858]

Height 	0.256 [0.096, 0.416] 	-0.045 [-0.103, 0.012]

Volume 	 	0.195 [0.173, 0.218]

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models,
  estimate = c("{estimate}{stars}",
               "{estimate} ({std.error})"))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816 (1.973)

(5.960) 	(1.973)

Height 	0.256** 	-0.045 (0.028)

(0.078) 	(0.028)

Volume 	 	0.195 (0.011)

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# vcov
modelsummary(models, vcov = "robust")

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(6.717) 	(2.544)

Height 	0.256 	-0.045

(0.090) 	(0.035)

Volume 	 	0.195

(0.014)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	8.021 	100.427

RMSE 	2.64 	0.75

Std.Errors 	HC3 	HC3

modelsummary(models, vcov = list("classical", "stata"))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(2.261)

Height 	0.256 	-0.045

(0.078) 	(0.031)

Volume 	 	0.195

(0.012)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	151.435

RMSE 	2.64 	0.75

Std.Errors 	IID 	HC1

modelsummary(models, vcov = sandwich::vcovHC)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(6.717) 	(2.544)

Height 	0.256 	-0.045

(0.090) 	(0.035)

Volume 	 	0.195

(0.014)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

RMSE 	2.64 	0.75

Std.Errors 	Custom 	Custom

modelsummary(models,
  vcov = list(stats::vcov, sandwich::vcovHC))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(2.544)

Height 	0.256 	-0.045

(0.078) 	(0.035)

Volume 	 	0.195

(0.014)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

RMSE 	2.64 	0.75

Std.Errors 	Custom 	Custom

modelsummary(models,
  vcov = list(c("(Intercept)"="", "Height"="!"),
              c("(Intercept)"="", "Height"="!", "Volume"="!!")))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

Height 	0.256 	-0.045

! 	!

Volume 	 	0.195

!!

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

RMSE 	2.64 	0.75

Std.Errors 	Custom 	Custom

# vcov with custom names
modelsummary(
  models,
  vcov = list("Stata Corp" = "stata",
              "Newey Lewis & the News" = "NeweyWest"))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(6.171) 	(1.839)

Height 	0.256 	-0.045

(0.083) 	(0.022)

Volume 	 	0.195

(0.007)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

RMSE 	2.64 	0.75

Std.Errors 	Stata Corp 	Newey Lewis &amp; the News

# fmt
mod <- lm(mpg ~ hp + drat + qsec, data = mtcars)

modelsummary(mod, fmt = 3)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	17.737

(13.020)

hp 	-0.058

(0.014)

drat 	4.429

(1.292)

qsec 	-0.284

(0.489)

Num.Obs. 	32

R2 	0.744

R2 Adj. 	0.717

AIC 	171.1

BIC 	178.5

Log.Lik. 	-80.561

F 	27.161

RMSE 	3.00

modelsummary(mod, fmt = fmt_significant(3))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	17.7

(13.0)

hp 	-0.0580

(0.0142)

drat 	4.43

(1.29)

qsec 	-0.284

(0.489)

Num.Obs. 	32

R2 	0.744

R2 Adj. 	0.717

AIC 	171.1

BIC 	178.5

Log.Lik. 	-80.561

F 	27.161

RMSE 	3.00

modelsummary(mod, fmt = NULL)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	17.7366200480466

(13.0197907534221)

hp 	-0.057972987581075

(0.014208609097443)

drat 	4.42875344653986

(1.29169225236228)

qsec 	-0.284074290906574

(0.489229280152)

Num.Obs. 	32

R2 	0.744

R2 Adj. 	0.717

AIC 	171.1

BIC 	178.5

Log.Lik. 	-80.561

F 	27.161

RMSE 	3.00

modelsummary(mod, fmt = fmt_decimal(4))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	17.7366

(13.0198)

hp 	-0.0580

(0.0142)

drat 	4.4288

(1.2917)

qsec 	-0.2841

(0.4892)

Num.Obs. 	32

R2 	0.744

R2 Adj. 	0.717

AIC 	171.1

BIC 	178.5

Log.Lik. 	-80.561

F 	27.161

RMSE 	3.00

modelsummary(mod, fmt = fmt_sprintf("%.5f"))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	17.73662

(13.01979)

hp 	-0.05797

(0.01421)

drat 	4.42875

(1.29169)

qsec 	-0.28407

(0.48923)

Num.Obs. 	32

R2 	0.744

R2 Adj. 	0.717

AIC 	171.1

BIC 	178.5

Log.Lik. 	-80.561

F 	27.161

RMSE 	3.00

modelsummary(mod, fmt = fmt_statistic(estimate = 4, conf.int = 1), statistic = "conf.int")

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	17.7366

[-8.9, 44.4]

hp 	-0.0580

[-0.1, 0.0]

drat 	4.4288

[1.8, 7.1]

qsec 	-0.2841

[-1.3, 0.7]

Num.Obs. 	32

R2 	0.744

R2 Adj. 	0.717

AIC 	171.1

BIC 	178.5

Log.Lik. 	-80.561

F 	27.161

RMSE 	3.00

modelsummary(mod, fmt = fmt_term(hp = 4, drat = 1, default = 2))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	17.74

(13.02)

hp 	-0.0580

(0.0142)

drat 	4.4

(1.3)

qsec 	-0.28

(0.49)

Num.Obs. 	32

R2 	0.744

R2 Adj. 	0.717

AIC 	171.1

BIC 	178.5

Log.Lik. 	-80.561

F 	27.161

RMSE 	3.00

m <- lm(mpg ~ I(hp * 1000) + drat, data = mtcars)
f <- function(x) format(x, digits = 3, nsmall = 2, scientific = FALSE, trim = TRUE)
modelsummary(m, fmt = f, gof_map = NA)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	10.7898612

(5.07775160)

I(hp * 1000) 	-0.0000518

(0.00000929)

drat 	4.6981578

(1.19163348)

# coef_rename
modelsummary(models, coef_rename = c('Volume' = 'Large', 'Height' = 'Tall'))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Tall 	0.256 	-0.045

(0.078) 	(0.028)

Large 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models, coef_rename = toupper)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(INTERCEPT) 	-6.188 	10.816

(5.960) 	(1.973)

HEIGHT 	0.256 	-0.045

(0.078) 	(0.028)

VOLUME 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models, coef_rename = coef_rename)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Volume 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# coef_rename = TRUE for variable labels
datlab <- mtcars
datlab$cyl <- factor(datlab$cyl)
attr(datlab$hp, "label") <- "Horsepower"
attr(datlab$cyl, "label") <- "Cylinders"
modlab <- lm(mpg ~ hp * drat + cyl, data = datlab)
modelsummary(modlab, coef_rename = TRUE)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	14.850

(11.923)

Horsepower 	-0.011

(0.082)

drat 	3.612

(3.019)

Cylinders [6] 	-4.091

(1.881)

Cylinders [8] 	-5.019

(3.056)

Horsepower × drat 	-0.006

(0.021)

Num.Obs. 	32

R2 	0.784

R2 Adj. 	0.742

AIC 	169.7

BIC 	180.0

Log.Lik. 	-77.860

RMSE 	2.76

# coef_rename: unnamed vector of length equal to the number of terms in the final table
m <- lm(hp ~ mpg + factor(cyl), data = mtcars)
modelsummary(m, coef_omit = -(3:4), coef_rename = c("Cyl 6", "Cyl 8"))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

Cyl 6 	16.623

(23.197)

Cyl 8 	88.105

(28.819)

Num.Obs. 	32

R2 	0.737

R2 Adj. 	0.709

AIC 	327.7

BIC 	335.0

Log.Lik. 	-158.831

F 	26.122

RMSE 	34.62

# coef_map
modelsummary(models, coef_map = c('Volume' = 'Large', 'Height' = 'Tall'))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

Large 	 	0.195

(0.011)

Tall 	0.256 	-0.045

(0.078) 	(0.028)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

modelsummary(models, coef_map = c('Volume', 'Height'))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

Volume 	 	0.195

(0.011)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# coef_omit: omit the first and second coefficients
modelsummary(models, coef_omit = 1:2)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

Volume 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# coef_omit: omit coefficients matching one substring
modelsummary(models, coef_omit = "ei", gof_omit = ".*")

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Volume 	 	0.195

(0.011)

# coef_omit: omit a specific coefficient
modelsummary(models, coef_omit = "^Volume$", gof_omit = ".*")

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

# coef_omit: omit coefficients matching either one of two substring
#modelsummary(models, coef_omit = "ei|rc", gof_omit = ".*")

# coef_omit: keep coefficients starting with a substring (using a negative lookahead)
#modelsummary(models, coef_omit = "^(?!Vol)", gof_omit = ".*")

# coef_omit: keep coefficients matching a substring
modelsummary(models, coef_omit = "^(?!.*ei|.*pt)", gof_omit = ".*")

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

# shape: multinomial model
library(nnet)
multi <- multinom(factor(cyl) ~ mpg + hp, data = mtcars, trace = FALSE) 

# shape: term names and group ids in rows, models in columns
modelsummary(multi, shape = response ~ model)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="response">response</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 	6 	0.500

6 	(41.760)

8 	8.400

8 	(0.502)

mpg 	6 	-83.069

6 	(416.777)

8 	-120.167

8 	(508.775)

hp 	6 	16.230

6 	(81.808)

8 	20.307

8 	(87.777)

Num.Obs. 	 	32

R2 	 	1.000

R2 Adj. 	 	0.971

AIC 	 	12.0

BIC 	 	20.8

RMSE 	 	0.00

# shape: term names and group ids in rows in a single column
modelsummary(multi, shape = term : response ~ model)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="(1)">(1)</th>
</tr>

(Intercept) 6 	0.500

(41.760)

mpg 6 	-83.069

(416.777)

hp 6 	16.230

(81.808)

(Intercept) 8 	8.400

(0.502)

mpg 8 	-120.167

(508.775)

hp 8 	20.307

(87.777)

Num.Obs. 	32

R2 	1.000

R2 Adj. 	0.971

AIC 	12.0

BIC 	20.8

RMSE 	0.00

# shape: term names in rows and group ids in columns
modelsummary(multi, shape = term ~ response:model)

<tr class="gt_col_headings gt_spanner_row">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="2" colspan="1" scope="col" id=" "> </th>
  <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="(1)    ">
    <span class="gt_column_spanner">(1)    </span>
  </th>
</tr>
<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="6">6</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="8">8</th>
</tr>

(Intercept) 	0.500 	8.400

(41.760) 	(0.502)

mpg 	-83.069 	-120.167

(416.777) 	(508.775)

hp 	16.230 	20.307

(81.808) 	(87.777)

Num.Obs. 	32

R2 	1.000

R2 Adj. 	0.971

AIC 	12.0

BIC 	20.8

RMSE 	0.00

# shape = "rcollapse"
panels <- list(
    "Panel A: MPG" = list(
        "A" = lm(mpg ~ hp, data = mtcars),
        "B" = lm(mpg ~ hp + factor(gear), data = mtcars)),
    "Panel B: Displacement" = list(
        "A" = lm(disp ~ hp, data = mtcars),
        "C" = lm(disp ~ hp + factor(gear), data = mtcars))
)

modelsummary(
    panels,
    shape = "rbind",
    gof_map = c("nobs", "r.squared"))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="A">A</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="B">B</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="C">C</th>
</tr>

Panel A: MPG

(Intercept) 	30.099 	27.882

(1.634) 	(2.109)

hp 	-0.068 	-0.067

(0.010) 	(0.011)

factor(gear)4 	 	2.635

(1.552)

factor(gear)5 	 	6.575

(1.643)

Num.Obs. 	32 	32

R2 	0.602 	0.753

Panel B: Displacement

(Intercept) 	20.992 	 	112.093

(32.607) 	 	(34.704)

hp 	1.430 	 	1.216

(0.202) 	 	(0.182)

factor(gear)4 	 	 	-97.923

(25.532)

factor(gear)5 	 	 	-147.495

(27.030)

Num.Obs. 	32 	 	32

R2 	0.626 	 	0.842

# title
modelsummary(models, title = 'This is the title')

This is the title

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Volume 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# title with LaTeX label (for numbering and referencing)
modelsummary(models, title = 'This is the title \\label{tab:description}')

This is the title \label{tab:description}

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Volume 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# add_rows
rows <- tibble::tribble(~term, ~Bivariate, ~Multivariate,
  'Empty row', '-', '-',
  'Another empty row', '?', '?')
attr(rows, 'position') <- c(1, 3)
modelsummary(models, add_rows = rows)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

Empty row 	- 	-

(Intercept) 	-6.188 	10.816

Another empty row 	? 	?

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Volume 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

# notes
modelsummary(models, notes = list('A first note', 'A second note'))

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Volume 	 	0.195

(0.011)

Num.Obs. 	31 	31

R2 	0.270 	0.941

R2 Adj. 	0.244 	0.937

AIC 	154.1 	78.2

BIC 	158.4 	84.0

Log.Lik. 	-74.061 	-35.116

F 	10.707 	222.471

RMSE 	2.64 	0.75

A first note

A second note

# gof_map: tribble
library(tibble)
gm <- tribble(
  ~raw,        ~clean,      ~fmt,
  "r.squared", "R Squared", 5)
modelsummary(models, gof_map = gm)

<tr class="gt_col_headings">
  <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id=" "> </th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Bivariate">Bivariate</th>
  <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="Multivariate">Multivariate</th>
</tr>

(Intercept) 	-6.188 	10.816

(5.960) 	(1.973)

Height 	0.256 	-0.045

(0.078) 	(0.028)

Volume 	 	0.195

(0.011)

R Squared 	0.26965 	0.94080

References

Arel-Bundock V (2022). “modelsummary: Data and Model Summaries in R.” Journal of Statistical Software, 103(1), 1-23. doi:10.18637/jss.v103.i01 .'